Multi-level Monte Carlo estimation test suite

Computes a suite of diagnostic values for an MLMC estimation problem.

Usage

mlmc.test(
  mlmc_l,
  N,
  L,
  N0,
  eps.v,
  Lmin,
  Lmax,
  parallel = NA,
  silent = FALSE,
  ...
)

Arguments

mlmc_l

a user supplied function which provides the estimate for level \(l\). It must take at least two arguments, the first is the level number to be simulated and the second the number of paths. Additional arguments can be taken if desired: all additional ... arguments to this function are forwarded to the user defined mlmc_l function.

The user supplied function should return a named list containing one element named sums and second named cost, where:

sums

is a vector of length two \(\left(\sum Y_i, \sum Y_i^2\right)\) where \(Y_i\) are iid simulations with expectation \(E[P_0]\) when \(l=0\) and expectation \(E[P_l-P_{l-1}]\) when \(l>0\).

cost

is a scalar with the total cost of the paths simulated. For example, in the financial options samplers included in this package, this is calculated as \(NM^l\), where \(N\) is the number of paths requested in the call to the user function mlmc_l, \(M\) is the refinement cost factor (\(M=2\) for mcqmc06_l() and \(M=4\) for opre_l()), and \(l\) is the level being sampled.

See the function (and source code of) opre_l() and mcqmc06_l() in this package for an example of user supplied level samplers.

N

number of samples to use in the tests

L

number of levels to use in the tests

N0

initial number of samples which are used for the first 3 levels and for any subsequent levels which are automatically added. Must be \(> 0\).

eps.v

a vector of one or more target accuracies for the tests. Must all be \(> 0\).

Lmin

the minimum level of refinement. Must be \(\ge 2\).

Lmax

the maximum level of refinement. Must be \(\ge\) Lmin.

parallel

if an integer is supplied, R will fork parallel parallel processes. This is done for the convergence tests section by splitting the N samples as evenly as possible across cores when sampling each level. This is also done for the MLMC complexity tests by passing the parallel argument on to the mlmc() driver when targeting each accuracy level in eps.

silent

set to TRUE to supress running output (identical output can still be printed by printing the return result)

...

additional arguments which are passed on when the user supplied mlmc_l function is called

Value

An mlmc.test object which contains all the computed diagnostic values. This object can be printed or plotted (see plot.mlmc.test).

Details

See one of the example level sampler functions (e.g. opre_l) for example usage.

This function is based on GPL-2 'Matlab' code by Mike Giles.

Author

Louis Aslett <louis.aslett@durham.ac.uk>

Mike Giles <Mike.Giles@maths.ox.ac.uk>

Tigran Nagapetyan <nagapetyan@stats.ox.ac.uk>

Examples

if (FALSE) { # \dontrun{
# Example calls with realistic arguments
# Financial options using an Euler-Maruyama discretisation
tst <- mlmc.test(opre_l, N = 2000000,
                 L = 5, N0 = 1000,
                 eps.v = c(0.005, 0.01, 0.02, 0.05, 0.1),
                 Lmin = 2, Lmax = 6,
                 option = 1)
tst
plot(tst)

# Financial options using a Milstein discretisation
tst <- mlmc.test(mcqmc06_l, N = 20000,
                 L = 8, N0 = 200,
                 eps.v = c(0.005, 0.01, 0.02, 0.05, 0.1),
                 Lmin = 2, Lmax = 10,
                 option = 1)
tst
plot(tst)
} # }

# Toy versions for CRAN tests
tst <- mlmc.test(opre_l, N = 10000,
                 L = 5, N0 = 1000,
                 eps.v = c(0.025, 0.1),
                 Lmin = 2, Lmax = 6,
                 option = 1)
#> 
#> **********************************************************
#> *** Convergence tests, kurtosis, telescoping sum check ***
#> *** using N = 10000 samples                            ***
#> **********************************************************
#> 
#>  l   ave(Pf-Pc)    ave(Pf)   var(Pf-Pc)    var(Pf)    kurtosis      check       cost
#> ---------------------------------------------------------------------------------------
#>  0   9.9864e+00  9.9864e+00  1.5806e+02  1.5806e+02  0.0000e+00  0.0000e+00  1.0000e+00 
#>  1   2.3616e-01  1.0528e+01  4.6195e+00  2.0513e+02  2.0180e+01  3.5003e-01  4.0000e+00 
#>  2   3.6916e-02  1.0709e+01  1.1293e+00  2.2145e+02  1.3114e+01  1.5936e-01  1.6000e+01 
#>  3   3.7534e-03  1.0384e+01  2.6563e-01  2.1298e+02  6.6822e+00  3.6542e-01  6.4000e+01 
#>  4   2.3320e-03  1.0483e+01  6.7543e-02  2.1957e+02  6.4185e+00  1.0816e-01  2.5600e+02 
#>  5   1.5741e-03  1.0830e+01  1.8175e-02  2.2166e+02  5.8218e+00  3.8607e-01  1.0240e+03 
#> 
#> ******************************************************
#> *** Linear regression estimates of MLMC parameters ***
#> ******************************************************
#> 
#>  alpha = 1.844283  (exponent for MLMC weak convergence)
#>  beta  = 2.004277  (exponent for MLMC variance) 
#>  gamma = 2.000000  (exponent for MLMC cost) 
#> 
#> ***************************** 
#> *** MLMC complexity tests *** 
#> ***************************** 
#> 
#>   eps      value    mlmc_cost   std_cost  savings     N_l 
#> ----------------------------------------------------------- 
#> 0.0250  1.0470e+01  1.370e+06  2.908e+07    21.22    685621     57426     13880      3639
#> 0.1000  1.0323e+01  6.383e+04  4.724e+05     7.40     35781      3013      1000
#> 

tst <- mlmc.test(mcqmc06_l, N = 10000,
                 L = 8, N0 = 1000,
                 eps.v = c(0.025, 0.1),
                 Lmin = 2, Lmax = 10,
                 option = 1)
#> 
#> **********************************************************
#> *** Convergence tests, kurtosis, telescoping sum check ***
#> *** using N = 10000 samples                            ***
#> **********************************************************
#> 
#>  l   ave(Pf-Pc)    ave(Pf)   var(Pf-Pc)    var(Pf)    kurtosis      check       cost
#> ---------------------------------------------------------------------------------------
#>  0   9.9254e+00  9.9254e+00  1.9093e+02  1.9093e+02  0.0000e+00  0.0000e+00  1.0000e+00 
#>  1   1.8237e-01  1.0161e+01  1.6474e-01  2.0577e+02  1.5619e+02  6.1665e-02  2.0000e+00 
#>  2   1.0610e-01  1.0567e+01  4.3677e-02  2.1657e+02  2.4251e+01  3.4217e-01  4.0000e+00 
#>  3   5.3915e-02  1.0397e+01  1.1391e-02  2.1350e+02  1.6232e+01  2.5413e-01  8.0000e+00 
#>  4   2.8379e-02  1.0610e+01  3.1579e-03  2.2062e+02  1.6977e+01  2.0880e-01  1.6000e+01 
#>  5   1.4092e-02  1.0527e+01  8.3114e-04  2.1503e+02  1.3656e+01  1.0984e-01  3.2000e+01 
#>  6   6.9390e-03  1.0364e+01  2.0637e-04  2.1144e+02  1.0092e+01  1.9322e-01  6.4000e+01 
#>  7   3.5004e-03  1.0411e+01  5.1394e-05  2.1552e+02  1.0496e+01  4.9106e-02  1.2800e+02 
#>  8   1.7754e-03  1.0504e+01  1.3201e-05  2.1573e+02  8.9490e+00  1.0384e-01  2.5600e+02 
#> 
#> ******************************************************
#> *** Linear regression estimates of MLMC parameters ***
#> ******************************************************
#> 
#>  alpha = 0.967513  (exponent for MLMC weak convergence)
#>  beta  = 1.942754  (exponent for MLMC variance) 
#>  gamma = 1.000000  (exponent for MLMC cost) 
#> 
#> ***************************** 
#> *** MLMC complexity tests *** 
#> ***************************** 
#> 
#>   eps      value    mlmc_cost   std_cost  savings     N_l 
#> ----------------------------------------------------------- 
#> 0.0250  1.0462e+01  5.305e+05  2.887e+07    54.42    471243      9357      3473      1309       468       158        57
#> 0.1000  1.0562e+01  3.641e+04  4.707e+05    12.93     29249      1000      1000        81        32
#>