Plot an mlmc.test object

Produces diagnostic plots on the result of an mlmc.test function call.

Usage

# S3 method for class 'mlmc.test'
plot(x, which = "all", cols = NA, ...)

Arguments

x

an mlmc.test object as produced by a call to the mlmc.test function.

which

a vector of strings specifying which plots to produce, or "all" to do all diagnostic plots. The options are:

"var" = \(log_2\) of variance against level;

"mean" = \(log_2\) of mean against level;

"consis" = consistency against level;

"kurt" = kurtosis against level;

"Nl" = \(log_2\) of number of samples against level;

"cost" = \(log_10\) of cost against \(log_10\) of epsilon (accuracy).

cols

the number of columns across to plot to override the default value.

...

additional arguments which are passed on to plotting functions.

Author

Louis Aslett <louis.aslett@durham.ac.uk>

Examples

if (FALSE) { # \dontrun{
tst <- mlmc.test(opre_l, N=2000000,
                 L=5, N0=1000,
                 eps.v=c(0.005, 0.01, 0.02, 0.05, 0.1),
                 Lmin=2, Lmax=6, option=1)
tst
plot(tst)
} # }