Multi-level Monte Carlo estimation test suite

Computes a suite of diagnostic values for an MLMC estimation problem.

Usage

mlmc.test(
  mlmc_l,
  N,
  L,
  N0,
  eps.v,
  Lmin,
  Lmax,
  parallel = NA,
  silent = FALSE,
  ...
)

Arguments

mlmc_l

a user supplied function which provides the estimate for level l

N

number of samples to use in the tests

L

number of levels to use in the tests

N0

initial number of samples which are used for the first 3 levels and for any subsequent levels which are automatically added. Must be \(> 0\).

eps.v

a vector of all the target accuracies in the tests. Must all be \(> 0\).

Lmin

the minimum level of refinement. Must be \(\ge 2\).

Lmax

the maximum level of refinement. Must be \(\ge\) Lmin.

parallel

if an integer is supplied, R will fork parallel parallel processes an compute each level estimate in parallel.

silent

set to TRUE to supress running output (identical output can still be printed by printing the return result)

...

additional arguments which are passed on when the user supplied mlmc_l function is called

Value

An mlmc.test object which contains all the computed diagnostic values. This object can be printed or plotted (see plot.mlmc.test).

Details

See one of the example level sampler functions (e.g. opre_l) for example usage.

This function is based on GPL-2 'Matlab' code by Mike Giles.

Author

Louis Aslett <louis.aslett@durham.ac.uk>

Mike Giles <Mike.Giles@maths.ox.ac.uk>

Tigran Nagapetyan <nagapetyan@stats.ox.ac.uk>

Examples

if (FALSE) { # \dontrun{
# Example calls with realistic arguments
tst <- mlmc.test(opre_l, N=2000000,
                 L=5, N0=1000,
                 eps.v=c(0.005, 0.01, 0.02, 0.05, 0.1),
                 Lmin=2, Lmax=6, option=1)
tst
plot(tst)

tst <- mlmc.test(mcqmc06_l, N=20000,
                 L=8, N0=200,
                 eps.v=c(0.005, 0.01, 0.02, 0.05, 0.1),
                 Lmin=2, Lmax=10, option=1)
tst
plot(tst)
} # }

# Toy versions for CRAN tests
tst <- mlmc.test(opre_l, N=10000,
                 L=5, N0=1000,
                 eps.v=c(0.025, 0.1),
                 Lmin=2, Lmax=6, option=1)
#> 
#> **********************************************************
#> *** Convergence tests, kurtosis, telescoping sum check ***
#> *** using N = 10000 samples                            ***
#> **********************************************************
#> 
#>  l   ave(Pf-Pc)    ave(Pf)   var(Pf-Pc)    var(Pf)    kurtosis      check       cost
#> ---------------------------------------------------------------------------------------
#>  0   1.0021e+01  1.0021e+01  1.5672e+02  1.5672e+02  0.0000e+00  0.0000e+00  1.0000e+00 
#>  1   2.1833e-01  1.0449e+01  4.6667e+00  2.0343e+02  2.4773e+01  2.4139e-01  4.0000e+00 
#>  2   1.8179e-02  1.0370e+01  1.0011e+00  2.1206e+02  9.9908e+00  1.0824e-01  1.6000e+01 
#>  3   2.2204e-03  1.0304e+01  2.7158e-01  2.0436e+02  7.8956e+00  7.7475e-02  6.4000e+01 
#>  4   2.4869e-03  1.0291e+01  6.5776e-02  2.1107e+02  6.1242e+00  1.7318e-02  2.5600e+02 
#>  5   1.8235e-03  1.0488e+01  1.7803e-02  2.2009e+02  5.9747e+00  2.2078e-01  1.0240e+03 
#> 
#> ******************************************************
#> *** Linear regression estimates of MLMC parameters ***
#> ******************************************************
#> 
#>  alpha = 1.667722  (exponent for MLMC weak convergence)
#>  beta  = 1.999615  (exponent for MLMC variance) 
#>  gamma = 2.000000  (exponent for MLMC cost) 
#> 
#> ***************************** 
#> *** MLMC complexity tests *** 
#> ***************************** 
#> 
#>   eps      value    mlmc_cost   std_cost  savings     N_l 
#> ----------------------------------------------------------- 
#> 0.0250  1.0447e+01  1.348e+06  2.790e+07    20.70    680294     57099     13650      3453
#> 0.1000  1.0444e+01  6.532e+04  4.524e+05     6.93     36378      3235      1000
#> 

tst <- mlmc.test(mcqmc06_l, N=10000,
                 L=8, N0=1000,
                 eps.v=c(0.025, 0.1),
                 Lmin=2, Lmax=10, option=1)
#> 
#> **********************************************************
#> *** Convergence tests, kurtosis, telescoping sum check ***
#> *** using N = 10000 samples                            ***
#> **********************************************************
#> 
#>  l   ave(Pf-Pc)    ave(Pf)   var(Pf-Pc)    var(Pf)    kurtosis      check       cost
#> ---------------------------------------------------------------------------------------
#>  0   9.9360e+00  9.9360e+00  1.9527e+02  1.9527e+02  0.0000e+00  0.0000e+00  1.0000e+00 
#>  1   1.8229e-01  1.0193e+01  1.4497e-01  2.0400e+02  3.6176e+01  8.7405e-02  2.0000e+00 
#>  2   1.0372e-01  1.0367e+01  4.3853e-02  2.1466e+02  4.5068e+01  8.0506e-02  4.0000e+00 
#>  3   5.4179e-02  1.0507e+01  1.2379e-02  2.2235e+02  2.0586e+01  9.5376e-02  8.0000e+00 
#>  4   2.7786e-02  1.0437e+01  3.1113e-03  2.1664e+02  1.4431e+01  1.0887e-01  1.6000e+01 
#>  5   1.3935e-02  1.0482e+01  8.0288e-04  2.1402e+02  1.2392e+01  3.4561e-02  3.2000e+01 
#>  6   7.4023e-03  1.0784e+01  2.1957e-04  2.2487e+02  9.9664e+00  3.3208e-01  6.4000e+01 
#>  7   3.4586e-03  1.0411e+01  5.1474e-05  2.1342e+02  9.0658e+00  4.2431e-01  1.2800e+02 
#>  8   1.7082e-03  1.0161e+01  1.2897e-05  2.0747e+02  9.8193e+00  2.8939e-01  2.5600e+02 
#> 
#> ******************************************************
#> *** Linear regression estimates of MLMC parameters ***
#> ******************************************************
#> 
#>  alpha = 0.967932  (exponent for MLMC weak convergence)
#>  beta  = 1.931833  (exponent for MLMC variance) 
#>  gamma = 1.000000  (exponent for MLMC cost) 
#> 
#> ***************************** 
#> *** MLMC complexity tests *** 
#> ***************************** 
#> 
#>   eps      value    mlmc_cost   std_cost  savings     N_l 
#> ----------------------------------------------------------- 
#> 0.0250  1.0438e+01  5.452e+05  3.070e+07    56.32    484421      9121      3530      1350       499       178        61
#> 0.1000  1.0320e+01  3.548e+04  4.622e+05    13.03     28515      1000      1000        70        25
#>